+++ eng/tools/lint-diff/test/generateReport.test.ts
@@ -1,0 +1,93 @@
+import { test, describe } from "vitest";
+
+import { AutorestRunResult } from "../src/util.js";
+import { getLintDiffViolations, arrayIsEqual } from "../src/generateReport.js";
+
+describe("getLintDiffViolations", async () => {
+  function createRunResult(stdout: string, stderr: string = ""): AutorestRunResult {
+    return {
+      rootPath: "string",
+      readme: "string",
+      tag: "string",
+      error: null,
+      stdout: stdout,
+      stderr: stderr,
+    };
+  }
+
+  test.concurrent("returns an empty array on no interesting violations", async ({ expect }) => {
+    const runResult =
+      createRunResult(`{"pluginName":"spectral","extensionName":"@microsoft.azure/openapi-validator","level":"information","message":"spectralPluginFunc: Validating OpenAPI spec. TypeSpec-generated: true. Path: 'file:///home/djurek/azure-rest-api-specs/specification/codesigning/resource-manager/Microsoft.CodeSigning/stable/2025-03-30/codeSigningAccount.json'"}
+{"pluginName":"spectral","extensionName":"@microsoft.azure/openapi-validator","level":"information","message":"openapiValidatorPluginFunc: Return"}`);
+
+    const violations = await getLintDiffViolations(runResult);
+    expect(violations).toEqual([]);
+  });
+
+  test.concurrent("returns an error on an interesting violation", async ({ expect }) => {
+    const runResult =
+      createRunResult(`{"pluginName":"spectral","extensionName":"@microsoft.azure/openapi-validator","level":"information","message":"spectralPluginFunc: Validating OpenAPI spec. TypeSpec-generated: true. Path: 'file:///home/djurek/azure-rest-api-specs/specification/codesigning/resource-manager/Microsoft.CodeSigning/stable/2025-03-30/codeSigningAccount.json'"}
+{"pluginName":"spectral","extensionName":"@microsoft.azure/openapi-validator","level":"error","message":"Top level property names should not be repeated inside the properties bag for ARM resource 'CodeSigningAccount'. Properties [properties.sku] conflict with ARM top level properties. Please rename these.","code":"ArmResourcePropertiesBag","details":{"jsonpath":["definitions","CodeSigningAccount"],"validationCategory":"ARMViolation","providerNamespace":false,"resourceType":false,"range":{"start":{"line":1036,"column":27},"end":{"line":1051,"column":6}}},"source":[{"document":"file:///home/djurek/azure-rest-api-specs/specification/codesigning/resource-manager/Microsoft.CodeSigning/stable/2025-03-30/codeSigningAccount.json","position":{"line":1036,"column":5}}]}
+{"pluginName":"spectral","extensionName":"@microsoft.azure/openapi-validator","level":"information","message":"openapiValidatorPluginFunc: Return"}`);
+
+    const violations = await getLintDiffViolations(runResult);
+    expect(violations.length).toEqual(1);
+    expect(violations[0].level).toEqual("error");
+    expect(violations[0].code).toEqual("ArmResourcePropertiesBag");
+  });
+
+  test.concurrent(
+    "returns an empty array on violations that don't have extensionname @microsoft.azure/openapi-validator",
+    async ({ expect }) => {
+      const runResult =
+        createRunResult(`{"pluginName":"spectral","extensionName":"@microsoft.azure/openapi-validator","level":"information","message":"spectralPluginFunc: Validating OpenAPI spec. TypeSpec-generated: true. Path: 'file:///home/djurek/azure-rest-api-specs/specification/codesigning/resource-manager/Microsoft.CodeSigning/stable/2025-03-30/codeSigningAccount.json'"}
+{"pluginName":"spectral","extensionName":"THIS IS FILTERED OUT","level":"error","message":"Top level property names should not be repeated inside the properties bag for ARM resource 'CodeSigningAccount'. Properties [properties.sku] conflict with ARM top level properties. Please rename these.","code":"ArmResourcePropertiesBag","details":{"jsonpath":["definitions","CodeSigningAccount"],"validationCategory":"ARMViolation","providerNamespace":false,"resourceType":false,"range":{"start":{"line":1036,"column":27},"end":{"line":1051,"column":6}}},"source":[{"document":"file:///home/djurek/azure-rest-api-specs/specification/codesigning/resource-manager/Microsoft.CodeSigning/stable/2025-03-30/codeSigningAccount.json","position":{"line":1036,"column":5}}]}
+{"pluginName":"spectral","extensionName":"@microsoft.azure/openapi-validator","level":"information","message":"openapiValidatorPluginFunc: Return"}`);
+
+      const violations = await getLintDiffViolations(runResult);
+      expect(violations).toEqual([]);
+    },
+  );
+});
+
+describe("arrayIsEqual", () => {
+  test.concurrent("returns true for equal arrays", async ({ expect }) => {
+    const a = ["a", "b", "c"];
+    const b = ["a", "b", "c"];
+
+    const result = arrayIsEqual(a, b);
+    expect(result).toEqual(true);
+  });
+
+  test.concurrent("returns false for different arrays", async ({ expect }) => {
+    const a = ["a", "b", "c"];
+    const b = ["a", "b", "d"];
+
+    const result = arrayIsEqual(a, b);
+    expect(result).toEqual(false);
+  });
+
+  test.concurrent("returns false for different lengths", async ({ expect }) => {
+    const a = ["a", "b", "c"];
+    const b = ["a", "b"];
+
+    const result = arrayIsEqual(a, b);
+    expect(result).toEqual(false);
+  });
+
+  test.concurrent("returns true for empty arrays", async ({ expect }) => {
+    const a: string[] = [];
+    const b: string[] = [];
+
+    const result = arrayIsEqual(a, b);
+    expect(result).toEqual(true);
+  });
+
+  test.concurrent("returns true for equal arrays with different types", async ({ expect }) => {
+    const a = ["a", 1, "c"];
+    const b = ["a", 1, "c"];
+
+    const result = arrayIsEqual(a, b);
+    expect(result).toEqual(true);
+  });
+});
