+++ eng/tools/lint-diff/test/processChanges.fs.test.ts
@@ -1,4 +1,4 @@
-import { getAffectedReadmes } from "../src/util.js";
+import { getAffectedReadmes } from "../src/processChanges.js";
 import { vol } from "memfs";
 import { beforeEach, vi, test, describe } from "vitest";
