+++ eng/tools/lint-diff/tsconfig.json
@@ -2,6 +2,9 @@
   "extends": "../tsconfig.json",
   "compilerOptions": {
     "outDir": "./dist",
-    "noImplicitReturns": true
+    "noImplicitReturns": true,
+    "allowJs": true,
+    "checkJs": true,
+
   }
 }
